template/{{cookiecutter.api_name}}/pyproject.toml
[project]
name = "data-go-mcp.{{ cookiecutter.api_name }}"
version = "{{ cookiecutter.version }}"
description = "MCP server for {{ cookiecutter.api_description }}"
readme = "README.md"
authors = [
    { name = "{{ cookiecutter.author_name }}" }
]
requires-python = ">={{ cookiecutter.python_version }}"
license = {text = "Apache-2.0"}
keywords = ["mcp", "korea", "api", "data.go.kr", "{{ cookiecutter.api_name }}"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.13.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]

[project.urls]
Homepage = "https://github.com/{{ cookiecutter.github_username }}/data-go-mcp-servers"
Repository = "https://github.com/{{ cookiecutter.github_username }}/data-go-mcp-servers"
Documentation = "https://github.com/{{ cookiecutter.github_username }}/data-go-mcp-servers/tree/main/src/{{ cookiecutter.api_name }}"
"Bug Tracker" = "https://github.com/{{ cookiecutter.github_username }}/data-go-mcp-servers/issues"
Changelog = "https://github.com/{{ cookiecutter.github_username }}/data-go-mcp-servers/blob/main/src/{{ cookiecutter.api_name }}/CHANGELOG.md"

[project.scripts]
"data-go-mcp.{{ cookiecutter.api_name }}" = "data_go_mcp.{{ cookiecutter.api_name_underscore }}.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["data_go_mcp"]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_default_fixture_loop_scope = "function"